Refactor directory structure and fix drift
