+++ .github/workflows/check_generated_code_drift.yml
@@ -3,13 +3,9 @@ on:
   push:
     branches:
       - main
-    paths:
-      - 'resources/services/**/*'
   pull_request:
     branches:
       - main
-    paths:
-      - 'resources/services/**/*'
 jobs:
   check_generated_code_drift:
     name: Check Generated Code for Drift
@@ -49,6 +45,10 @@ jobs:
           key: ${{ runner.os }}-go-${{ matrix.go }}-${{ hashFiles('**/go.sum') }}
           restore-keys: |
             ${{ runner.os }}-go-${{ matrix.go }}-
+      - name: Fail if new cq-gen config file is missing //check-for-changes
+        if: steps.changes.outputs.src == 'true' || github.event_name != 'pull_request'
+        run: |
+          ./scripts/check-new-files-have-check-for-changes-flag.sh
       - name: Run go generate on changed service directories
         if: steps.changes.outputs.src == 'true' || github.event_name != 'pull_request'
         run: |
